backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-import-export==4.3.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
gunicorn==23.0.0
pillow==11.0.0
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
whitenoise==6.6.0